Adapt to RolloutModelConfig
